Выносит основную логику из main.js в контроллер
